update command name
